rename: chatMessage package 생성 후 로직 위치 변경
