+++ src/main/java/com/lawProject/SSL/domain/chatmessage/model/SenderType.java
@@ -1,0 +1,5 @@
+package com.lawProject.SSL.domain.chatmessage.model;
+
+public enum SenderType {
+    USER, AI
+}
