--- src/main/java/com/lawProject/SSL/domain/langchain/domain/SenderType.java
@@ -1,5 +1,0 @@
-package com.lawProject.SSL.domain.langchain.domain;
-
-public enum SenderType {
-    USER, AI
-}
